Use active file's project folder
Currently if config.startingDir is 'project' and multiple project
folders are open in the current window, the root of the project at the
top of the list when first launched is always used.

If the user edits a file from another project, this may not be
desirable. If the file is part of a Node.js package, for example, the
user may need to run Node Package Manager (npm) commands from the
file's project root, and similarly for git commands. This patch
updates the 'project' setting to check the project root of the
currently open file, falling back to the old behavior if no such file
is open.

Since the active file can change, we re-evaluate the starting directory
each time a shell process is spawned instead of permanently storing it
in @options.cwd with the existential assignment operator. That way, we
can open a file and then terminate our shell in order to open a fresh
shell in the new file's project root.